preprocessing_service/requirements_preprocessing.txt
numpy
fastapi
python-dotenv